requirements.txt
##################################################################
# core python modules
##################################################################
Django==1.11.1

##################################################################
# heroku hosting
##################################################################
dj-database-url==0.4.1
whitenoise==3.2
psycopg2
django_redis

#
uwsgi
django_cors_headers==2.1.0

##################################################################
# django modules
##################################################################
django-braces==1.10.0
django-celery-beat==1.0.1
django-click==1.2.0
django-compressor==2.0
django-split-settings==0.2.3

##################################################################
# RESTful API
# http://www.django-rest-framework.org/#installation
##################################################################
djangorestframework==3.6.3
markdown==2.6.7
coreapi==2.1.1

##################################################################
# authentication
##################################################################
django-authtools==1.5.0

##################################################################
# task handling
##################################################################
celery[redis]==4.0.2

##################################################################
# other modules
##################################################################
raven==6.0.0

##################################################################
# external modules & patches
##################################################################
-e git+https://github.com/palmbeach-interactive/django-slick-admin@master#egg=django-slick-admin
